+++ tests/Fhaculty/Graph/GraphTest.php
@@ -4,6 +4,7 @@ use Fhaculty\Graph\Exception\RuntimeException;
 use Fhaculty\Graph\Exporter\Image;
 use Fhaculty\Graph\Vertex;
 use Fhaculty\Graph\Exception\OverflowException;
+use Fhaculty\Graph\Exception\InvalidArgumentException;
 use Fhaculty\Graph\Graph;
 
 class GraphTest extends TestCase
@@ -177,5 +178,79 @@ class GraphTest extends TestCase
 
         $this->assertEquals(array(1, 3), array_keys($v2->getVerticesEdgeTo()));
         $this->assertEquals(array(1), array_keys($v2->getVerticesEdgeFrom()));
+    }
+
+    public function testCreateVerticesNone()
+    {
+        $graph = new Graph();
+
+        $this->assertEquals(array(), $graph->createVertices(0));
+        $this->assertEquals(array(), $graph->createVertices(array()));
+
+        $this->assertEquals(0, $graph->getNumberOfVertices());
+    }
+
+    /**
+     * expect to fail for invalid number of vertices
+     * @expectedException InvalidArgumentException
+     * @dataProvider testCreateVerticesFailProvider
+     */
+    public function testCreateVerticesFail($number)
+    {
+        $graph = new Graph();
+        $graph->createVertices($number);
+    }
+
+    public static function testCreateVerticesFailProvider()
+    {
+        return array(
+            array(-1),
+            array("10"),
+            array(0.5),
+            array(null),
+            array(array(1, 1))
+        );
+    }
+
+    public function testCreateVerticesOkay()
+    {
+        $graph = new Graph();
+
+        $vertices = $graph->createVertices(2);
+        $this->assertCount(2, $vertices);
+        $this->assertEquals(array(0, 1), array_keys($graph->getVertices()));
+
+        $vertices = $graph->createVertices(array(7, 9));
+        $this->assertCount(2, $vertices);
+        $this->assertEquals(array(0, 1, 7, 9), array_keys($graph->getVertices()));
+
+        $vertices = $graph->createVertices(3);
+        $this->assertCount(3, $vertices);
+        $this->assertEquals(array(0, 1, 7, 9, 10, 11, 12), array_keys($graph->getVertices()));
+    }
+
+    public function testCreateVerticesAtomic()
+    {
+        $graph = new Graph();
+
+        // create vertices 10-19 (inclusive)
+        $vertices = $graph->createVertices(range(10, 19));
+        $this->assertCount(10, $vertices);
+
+        try {
+            $graph->createVertices(array(9, 19, 20));
+            $this->fail('Should be unable to create vertices because of duplicate IDs');
+        }
+        catch (OverflowException $ignoreExpected) {
+            $this->assertEquals(10, $graph->getNumberOfVertices());
+        }
+
+        try {
+            $graph->createVertices(array(20, 21, 21));
+            $this->fail('Should be unable to create vertices because of duplicate IDs');
+        }
+        catch (InvalidArgumentException $ignoreExpected) {
+            $this->assertEquals(10, $graph->getNumberOfVertices());
+        }
     }
 }